refactor: move interfaces to separate folder
