+++ src/components/Heading/Heading.tsx
@@ -1,6 +1,6 @@
 import './Heading.css';
 import React from 'react';
-import { IHeadingProps } from './IHeadingProps';
+import { IHeadingProps } from '../../interfaces/IHeadingProps';
 
 class Heading extends React.Component<IHeadingProps> {
   constructor(props: IHeadingProps) {
