+++ src/interfaces/ICardProps.ts
@@ -1,4 +1,4 @@
-import IProduct from '../../interfaces/IProduct';
+import IProduct from './IProduct';
 
 export default interface ICardProps {
   value: IProduct;
