+++ src/interfaces/IContentProps.ts
